Use shouldEqualColor in all tests

+++ src/commonTest/kotlin/com/github/ajalt/colormath/Matchers.kt
@@ -9,7 +9,6 @@ import io.kotest.matchers.Matcher
 import io.kotest.matchers.MatcherResult
 import io.kotest.matchers.floats.plusOrMinus
 import io.kotest.matchers.shouldBe
-import kotlin.math.abs
 
 fun convertTo(expected: Color) = object : Matcher<Color> {
     override fun test(value: Color): MatcherResult {
@@ -25,12 +24,12 @@ fun convertTo(expected: Color) = object : Matcher<Color> {
     }
 }
 
-fun Color.shouldEqualColor(expected: Color, tolerance: Float = 0.0005f) {
+fun Color.shouldEqualColor(expected: Color, tolerance: Double = 0.0005) {
     try {
         this::class shouldBe expected::class
         components().size shouldBe expected.components().size
         components().zip(expected.components()).forEach { (a, e) ->
-            a shouldBe (e plusOrMinus tolerance)
+            a shouldBe (e plusOrMinus tolerance.toFloat())
         }
     } catch (e: AssertionError) {
         println(this)
